part_1 (REST API)/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.1
SQLAlchemy==1.4.31
alembic==1.7.0
psycopg2-binary==2.9.1
pydantic==1.8.2
passlib==1.7.4
python-jose==3.3.0
gunicorn==20.0.4
